requirements.txt
# Production dependencies
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
umap-learn>=0.5.3
hdbscan>=0.8.30
shap>=0.42.0
lime>=0.2.0
plotly>=5.15.0
gradio>=4.0.0
pyyaml>=6.0.0